Pull toRemove step into BufferReflow
