+++ src/Buffer.ts
@@ -9,7 +9,7 @@ import { EventEmitter } from './common/EventEmitter';
 import { IMarker } from 'xterm';
 import { BufferLine } from './BufferLine';
 import { DEFAULT_COLOR } from './renderer/atlas/Types';
-import { reflowSmallerGetNewLineLengths } from './BufferReflow';
+import { reflowSmallerGetNewLineLengths, reflowLargerGetLinesToRemove } from './BufferReflow';
 
 export const DEFAULT_ATTR = (0 << 18) | (DEFAULT_COLOR << 9) | (256 << 0);
 export const CHAR_DATA_ATTR_INDEX = 0;
@@ -26,7 +26,7 @@ export const WHITESPACE_CELL_CHAR = ' ';
 export const WHITESPACE_CELL_WIDTH = 1;
 export const WHITESPACE_CELL_CODE = 32;
 
-const FILL_CHAR_DATA: CharData = [DEFAULT_ATTR, NULL_CELL_CHAR, NULL_CELL_WIDTH, NULL_CELL_CODE];
+export const FILL_CHAR_DATA: CharData = [DEFAULT_ATTR, NULL_CELL_CHAR, NULL_CELL_WIDTH, NULL_CELL_CODE];
 
 /**
  * This class represents a terminal buffer (an internal state of the terminal), where the
@@ -240,78 +240,11 @@ export class Buffer implements IBuffer {
   }
 
   private _reflowLarger(newCols: number): void {
+    // TODO: Can toRemove be pulled out into BufferReflow?
+
     // Gather all BufferLines that need to be removed from the Buffer here so that they can be
     // batched up and only committed once
-    const toRemove: number[] = [];
-    for (let y = 0; y < this.lines.length - 1; y++) {
-      // Check if this row is wrapped
-      let i = y;
-      let nextLine = this.lines.get(++i) as BufferLine;
-      if (!nextLine.isWrapped) {
-        continue;
-      }
-
-      // Check how many lines it's wrapped for
-      const wrappedLines: BufferLine[] = [this.lines.get(y) as BufferLine];
-      while (i < this.lines.length && nextLine.isWrapped) {
-        wrappedLines.push(nextLine);
-        nextLine = this.lines.get(++i) as BufferLine;
-      }
-
-      // Copy buffer data to new locations
-      let destLineIndex = 0;
-      let destCol = wrappedLines[destLineIndex].getTrimmedLength();
-      let srcLineIndex = 1;
-      let srcCol = 0;
-      while (srcLineIndex < wrappedLines.length) {
-        const srcTrimmedTineLength = wrappedLines[srcLineIndex].getTrimmedLength();
-        const srcRemainingCells = srcTrimmedTineLength - srcCol;
-        const destRemainingCells = newCols - destCol;
-        const cellsToCopy = Math.min(srcRemainingCells, destRemainingCells);
-
-        wrappedLines[destLineIndex].copyCellsFrom(wrappedLines[srcLineIndex], srcCol, destCol, cellsToCopy, false);
-
-        destCol += cellsToCopy;
-        if (destCol === newCols) {
-          destLineIndex++;
-          destCol = 0;
-        }
-        srcCol += cellsToCopy;
-        if (srcCol === srcTrimmedTineLength) {
-          srcLineIndex++;
-          srcCol = 0;
-        }
-
-        // Make sure the last cell isn't wide, if it is copy it to the current dest
-        if (destCol === 0) {
-          if (wrappedLines[destLineIndex - 1].getWidth(newCols - 1) === 2) {
-            wrappedLines[destLineIndex].copyCellsFrom(wrappedLines[destLineIndex - 1], newCols - 1, destCol++, 1, false);
-            // Null out the end of the last row
-            wrappedLines[destLineIndex - 1].set(newCols - 1, FILL_CHAR_DATA);
-          }
-        }
-      }
-
-      // Clear out remaining cells or fragments could remain;
-      wrappedLines[destLineIndex].replaceCells(destCol, newCols, FILL_CHAR_DATA);
-
-      // Work backwards and remove any rows at the end that only contain null cells
-      let countToRemove = 0;
-      for (let i = wrappedLines.length - 1; i > 0; i--) {
-        if (i > destLineIndex || wrappedLines[i].getTrimmedLength() === 0) {
-          countToRemove++;
-        } else {
-          break;
-        }
-      }
-
-      if (countToRemove > 0) {
-        toRemove.push(y + wrappedLines.length - countToRemove); // index
-        toRemove.push(countToRemove);
-      }
-
-      y += wrappedLines.length - 1;
-    }
+    const toRemove: number[] = reflowLargerGetLinesToRemove(this.lines, newCols);
 
     if (toRemove.length > 0) {
       // First iterate through the list and get the actual indexes to use for rows
